requirements.txt
pdf2image
pytesseract
Pillow
fpdf
fuzzywuzzy
python-Levenshtein
opencv-python-headless
numpy